DeepSeek-Agent/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
pillow>=10.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
pyyaml>=6.0
pydantic>=2.0.0
accelerate>=0.27.0
vllm>=0.3.0
playwright>=1.40.0
pyautogui>=0.9.53
easydict>=1.9
einops>=0.7.0
flash-attn>=2.5.0
peft>=0.14.0
